// gig-rdo-vite/src/types/patrimonio-types.ts
export interface PatrimonioListFilters {
  search?: string;
  categoria?: string;
  status?: 'todos' | 'ativo' | 'baixado';
  mobilizado?: 'todos' | 'sim' | 'nao';
  temPatrimonio?: 'todos' | 'sim' | 'nao';
  empresa?: number;
}